requirements.txt
Flask
cloudinary
python-dotenv
werkzeug
gunicorn
